Update/extend some of the CoC language
I tried to express something I'd feel comfortable personally saying.  In general, I'm concerned that codes of conduct set the bar at "no harassment" as compared to "actively empowering of all." It seems to me that microaggressions and biases that are not necessarily _harassment_ should be mentioned in a code of conduct. Please let me know if this in some way dilutes the effectiveness of the CoC